Stokhos: Adaptivity manager now correctly handles non-contiguous entries
I've been lucky for a while, but finally my luck ran out with the
new DOFManager in panzer which produces (even in serial) non-contiguous
GID entries for a map. This would have been a problem in parallel as
well. This change just makes sure the linear system assembled is the
correct one for the residual.