Ifpack2: Relaxation: Removed unnecessary state.
Relaxation was storing some fields that are inexpensive to recompute.
Some of them depend on the structure of the matrix and may change if
the matrix changes.  It's generally better just to ask the matrix each
time for that data, rather than to have to think about when and whether
the matrix might have changed each time you access cached data.

This commit removes all such data from Relaxation.  That has the side
benefit of making the initialize() method almost trivial.

Relaxation now uses TimeMonitor to manage starting and stopping the
timer, rather than doing it manually.  This is more exception safe.