Panzer: Fixed instantiation of initial condition builder
This object used to be templated.  When the DOFManager was removed as a parameter, then the template arguments were removed.  But this object we never changed to go from a _impl.hpp to a .cpp file.  We got away this this because it was only used in the Drekar_Main.cpp file (limited to one instatiation).  When I moved the guts of Drekar_Main into a driver class that is part of the drekar library, we ended up having multiple (4) instantiations of this object that collided during the final link.  This commit fixes the problem.