Panzer: Generalizing coordinate construction to work for either DOFManager
The code has been simplified and unified a bit so this now works
between DOFManagers. This is still a bit of a clunky part of the
code and probably won't work for blocked systems. This needs to
be fixed and instead I should use the response library to construct
all the coordinates. (The difficulty here is for mixed discretizations
in case you were wondering!).

Also added an accessor to the IntrepidFP to determine if it supports
coordinates or not.