Tpetra: Implemented removeEmptyProcessesInPlace for MV, graph, & matrix.
This commit adds implementations of DistObject's
removeEmptyProcessesInPlace to MultiVector, CrsGraph, and CrsMatrix.
The method replaces the object's Map (row Map in the case of CrsGraph
and CrsMatrix) with a new Map, from which processes with zero entries
have been removed.

Users may access this functionality through the nonmember functions
removeEmptyProcessesInPlace that take a DistObject, and defer to the
subclass' implementation of the removeEmptyProcessesInPlace method.