Stokhos: Fix issue caused by backwards compatibility changes in epetra
The changes to fix a backwards compatibility issue in Epetra caused an ambiguous
call in a Stokhos test. The issue was the test was passing in a size_t type when
epetra was expecting an int or long long. To solve this the size_t variables
were cast to ints.

Build/Test Cases Summary
Enabled Packages: Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=58,notpassed=0 (21.15 min)
3) SERIAL_RELEASE_SS => passed: passed=58,notpassed=0 (17.85 min)