Tpetra: MatrixMarket: Added readSparse(File) with input Maps.
Siva Rajamanickam pointed out that the Tpetra functions for reading in
a sparse matrix from a Matrix Market file or input stream were missing
the option to provide a previously computed (or read in) row Map.  I
added these two functions: one for reading from a file given a filename,
and the other for reading from an already open input stream.