Tpetra: MultiVector: made replaceMap work for {sub,super}set comms.
replaceMap only used to work if the input Map had a communicator with
the same number of processes as the MultiVector's current Map.  An
important use case for MueLu, however, is to "restrict" a MultiVector
to a subset communicator, and "project" it back onto the original
communicator again.  Rather than write a new method for this, I decided
to extend the behavior of replaceMap to work for MueLu's use cases.
This does not change the behavior of replaceMap for the previous use
case.