Tpetra: Map: Defer Directory creation until first use.
The Map's Directory object is only used in getRemoteIndexList().  That
method must be called collectively over all processes in the Map's
communicator, just like creation of the Directory.  Thus, we do not need
to create the Directory in Map's constructor; we can defer Directory
creation to (the first call to) getRemoteIndexList().

This optimization saves the communication cost of creating the
Directory, for some Maps which are never involved in an Import or Export
operation.  For example, a nonsquare sparse matrix (CrsMatrix) with
identical row and range Maps would never need to construct an Export
object.  This is a common case for the prolongation or restriction
operators in algebraic multigrid.  Thus, this commit may reduce MueLu's
setup time.  Note, however, that it does not make the Directory creation
time go away if one actually needs to create an Import or Export with
that Map; it merely shifts the Directory creation time away from Map's
constructor.

Thanks to Chris Siefert for pointing out this optimization.

Build/Test Cases Summary
Enabled Packages: Teuchos, Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=272,notpassed=0 (4.61 min)
3) SERIAL_RELEASE_SS => passed: passed=254,notpassed=0 (10.39 min)