MueLu: a new approach to Zoltan2 partitioning
The old approach was to try to present the number of partitions as a
product of three numbers. It did not work so well. For instance, if we
had a 1172 parts, we would get 2x2x293, which produced very stretched
subdomains. These kind of subdomains are bad particularly if we use
uncoupled aggregation, as the boundary to inner node ratio is high.

It turns out that pqJagged algorithm in Zoltan2 also supports the
recursive "bisection" somewhat similar to Zoltan. So the only problem is
to come up with a series of small numbers for these recursion. I
achieved that by tweaking the number of processors slightly (less than
15%) so that the final number is of form 2^k*3^m. And now we use the
recursive form of pqJagged.

But the depth of recursion also matters, as performance of Zoltan2
deteriorates with deep recursion (more than 6, really). Therefore, there
are some algorithms in place to aggregate factors. The current
implementation is way too complex for what it does, and might be a bit
finicky, but should work.