TriUtils: Fix a backwards compatibility issue
To support the 64bit epetra work a couple classes had an additional parameter
added to their constructor. This new parameter had a default value for every
case except when both 32bit and 64bit code was enabled which caused the older
calls to the constructor to fail. The solution was to set a default for this
case which is the same as the no 64bit case to preserve the older 32bit
functionality of these classes.

Build/Test Cases Summary
Enabled Packages: Triutils
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=785,notpassed=0 (57.35 min)
1) SERIAL_RELEASE => passed: passed=757,notpassed=0 (8.07 min)
Other local commits for this build/test group: cdbba49