Updated numpy.i to avoid deprecated code
The NumPy API is currently at version 1.7, which gives deprecation
warnings for certain functions and macros.  Version 1.8 will be out
soon, which will not accept this deprecated code.  This change to
numpy.i checks the NumPy version and uses the appropriate API.

Note that this does not yet make PyTrilinos compliant with the new
NumPy API.  At the very least, the Epetra_NumPy* classes have to be
updated, and probably some others.