Kokkos: DefaultArithmetic: Full specialization for double,SerialNode.
This commit adds a full specialization of DefaultArithmetic (which
implements local MultiVector operations) for Scalar=double and Node =
SerialNode.  We already had a partial specialization for Node =
SerialNode; this full specialization lets us implement critical kernels
like Assign in the fastest way possible for this special case.  (For
example, the full specialization can use memcpy for Assign, since it is
correct for copying an array of double, but not necessarily correct for
copying an array of general Scalar.)