Update release tarball configure script
the 1.2.2 version of mpi4py was not playing well with some changes to pytrilinos
that improved the way that mpi4py was found. This was likely due to it being
installed incorrectly. The solution was to install the most recent version
correctly and make a few changes to support the new way of finding mpi4py to the
configure script