Adding a dynamic strategy for choosing how many subspace vectors to keep during a restart. This strategy is similar to the one used by ARPACK and will only be employed for a block size of 1 when the user sets the new parameter:
MyPL.set( "Dynamic Extra NEV", true );

By default this option is false, resulting in the standard behavior.  At this time, this
option will work with the static strategy for choosing subspace vectors to keep during a
restart:

MyPL.set( "Extra NEV Blocks", <#> );

However, if this approach proves to be superior, the later strategy will be deprecated.