Thyra: Specialized contiguous Tpetra MV subviews
When contiguous views are requested, instances of the proper subtype
TpetraMultiVector are returned instead of DefaultColumnwiseMultiVector
objects created by the default implementation provided in
MultiVectorDefaultBase.

That change is required to make the Tpetra concrete adapter compatible
with Belos.

The implementation is directly inspired by the helper functions in
Thyra_TpetraThyraWrappers. It may not be robust or general enough.

The virtual functions corresponding to noncontiguous views have not been
overridden and therefore the default implementation provided in
MultiVectorDefaultBase is still used in that case.