Tpetra: Export: Optimized construction.
This commit applies the same optimization to Export, as did my
previous commit to Import.  See the log for the following commit:

9d4c1721d129d41945ab136c4173ca663f1b224e

I also made Import::setupSamePermuteRemote look as much as (the
corresponding part of) Export::setupSamePermuteExport as possible,
to make it easier to apply further optimizations to both methods.