Resurrect a file in the SIERRA directory
The current configure logic always updates the Trilinos_version.h file for
Sierra and the recent removal of that dir will make it so any configure will be
adding a new file potentially causing the checkin test script to complain of
modified files that aren't committed. This just readds the files that are
configured every time in the SIERRA directory.

at some point in the near future the version file in SIERRA should only be
updated when doing a configure for Sierra, but this is a decent workaround for
the short term.