Changing the SpmdVectorBase to use new SpmdLocalDataAccess functions (Trilinos #5902)
I had to deal with the issue of automatic compatibility of vector spaces and
product vectors spaces with one block containing that same vector space.

I added a new unit test to once and for all remove the assumption that
sinlge-process vector spaces are compatible just becuase they have the same
size.  That is a (nice but not critial) feature that is just too complex to
support.  Supporting that is too much like magic.