Tpetra: FixedHashTable: Fixed compile error with gcc 4.4.6.
I didn't see this with Clang 3.2 on Mac, but did see it with gcc 4.4.6
on Linux.  Not sure what's wrong, but the change fixes the issue.
I've seen this issue before with ArrayRCP's nonmember constructor arcp
in the implementation of Tpetra::DistObject (it even has comments to
that effect)

I am pushing this change directly without running the check-in test
script, because I want to get this in now before it breaks any of the
gcc 4.4.6 builds.