Tpetra: Map: isSameAs now tests for congruence of the communicators.
isSameAs no longer requires that the input Map's communicator is
identical to the left-hand side Map's communicator.  Instead, it now
requires that the two communicators are congruent (have the same
numbers of processes in the same order).  (We never tested that the
communicators were identical before.)  Making the test for congruence
explicit should help with debugging any issues related to
subcommunicators.  It should also make it easier for Tpetra to
duplicate communicators in the way that the MPI standard recommends
(to prevent pollution of the user's communicator with nonblocking
messages from Tpetra operations), even though we don't do this yet.