Simple STK::transfer test.
Filled in the outline created by Steve Kennon
with a very simple transfer example that will
be used to research algorithms.

The use case is linear interpolation from a
random set of point/values to another set
of points.  This is very simple with the
stk::search class that is already in place but
demonstrates many of the attributes that
are in complicated transfers:

(1) Framework-like user defined tolerances
used to define a search distance around
the points. How to automatically define
this rather obscure parameter?

(2) What to do when tolerance from (1) is
too small or too large? Leads to yet another
user defined multiplication factor.

(3) Incremental update of position. Is
there a way to use stk::search to optimize
the algorithm if locations are incrementally
updated?

(4) How to handle parallel communication?
Currently is only written for serial and
is independent of stk::Mesh.

(5) Scaling studies would be nice.

One thing that this example is not very
useful for is studying conservative transfers.
That will need to wait for a mesh based transfer
example.

It's been a while, hopefully this does not break
anything.

(From Sierra)