Tpetra: CrsGraph: Avoid recomputing local constants every fillComplete.
If a CrsMatrix has a nonconst CrsGraph, then the CrsMatrix's
fillComplete will invoke the CrsGraph's fillComplete, even if the
structure of the graph hasn't changed.  CrsGraph's fillComplete always
calls computeGlobalConstants.  That method in turn always recomputes
"local constants" (like the local number of diagonal entries), before
it uses all-reduces to compute "global constants" (like the global
total number of diagonal entries).  We can't avoid at least one
all-reduce, because we have to know if any process' graph's structure
has changed.  However, we can avoid recomputation of local constants
by keeping track of whether the graph's structure has changed since
the last time local constants were computed.  Computing local
constants requires iterating over all the local data (to count
diagonal elements and check for lower and upper triangularity, for
example), so it would be worthwhile to avoid this cost if the graph's
structure has not changed.

This commit adds this optimization.  It adds a Boolean field
haveLocalConstants_ to CrsGraph.  Any method that changes the
structure of the matrix sets this to false, just like it sets the
"indices are sorted" and "indices are nonredundant" fields
(conservatively) to false.

The commit also replaces the setSorted and setMerged methods of
CrsGraph with a single setLocallyModified method.  The name makes the
meaning of the method more clear.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=173,notpassed=0 (9.14 min)
3) SERIAL_RELEASE_SS => passed: passed=157,notpassed=0 (11.59 min)