Tpetra: RowMatrix: Added sparse matrix add method, with passing test.
RowMatrix now has a new virtual method add(alpha,A,beta,...).  When
called on a RowMatrix instance B, it computes C = alpha*A + beta*B,
returning a new CrsMatrix C as a RowMatrix.  It comes with a default
implementation, which subclasses (like CrsMatrix) may override, for
example to optimize for the subclass' special case, or to return a
RowMatrix which is an instance of a different subclass.  I plan to
implement a more optimized CrsMatrix version of this method.  The
RowMatrix method already optimizes for the common case that the two
input RowMatrix objects have the same row Map.

I also added a new Tpetra::MatrixMatrix::add method (note lowercase;
this differs from the existing Tpetra::MatrixMatrix::Add methods, on
purpose because it has a different interface).  This uses the
RowMatrix::add method.  I added a test for this case in the
Tpetra::MatrixMatrix unit tests; the test passes.

Both the new methods (RowMatrix::add and Tpetra::MatrixMatrix::add)
take optional domain and range Map arguments.  This lets the method
work even if the input matrices are not fill complete (and thus may
not necessarily have domain and range Maps).  They also take a
ParameterList to govern whether fillComplete is called on C, and the
parameters passed to C's constructor and to its fillComplete call (if
applicable).  The RowMatrix::add public documentation explains these
parameters.

This commit completes my work-around for Bug 5916 ("3-argument
MatrixMatrix::Add does not work for null C input").  I can't change
the interface of MatrixMatrix::Add (note initial capital), for reasons
of backwards compatibility.  However, I can (and do here) offer a new
method, MatrixMatrix::add, which implements the desired functionality.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=173,notpassed=0 (4.55 min)
3) SERIAL_RELEASE_SS => passed: passed=157,notpassed=0 (11.58 min)