First checkin for zapotec2
This includes all the modifications to get basic zapotec2 working.
This only enables hexes, and has run just a few tests. Plenty
more work to be done on it.

includes:
- updates to jamfiles to build new code
- updates to cth TPL to avoid initialization problems
- addition of external calling routines into adagio (in adagio src)
- forcing of adagio to compute faces (in adagio, Agio_Region.C)
- addition of user subroutine to enable cth force vector
- Extensive modifications of zapotec interfaces to support
	needed data
- updating of zapotec application code

A few fixes to zapotec2 -- better integration

Used to have to force adagio to provide all faces, now have enabled
that to only happen when it is zapotec. Did a few other little
fixes to support old zapotec test.

(From Sierra)