Tpetra: Attempt to fix recent build errors.
Some compilers don't like definitions of template functions that are
namespace qualified -- they prefer the namespace to be outside.  For
example:

namespace MyNamespace {
  template<class T>
  T foo (const T& x) {
    return x + x;
  }
}

works, but

tmeplate<class T>
T MyNamespace::foo (const T& x) {
  return x + x;
}

does not.  I've fixed one of the functions that appears to be causing
offense on a few different compilers (not gcc 4.4.6).  Not sure if
this will help the build errors, but it never hurts.