Tpetra: Directory: Optimized sparse Maps (Bug 5822) fix.
This commit optimizes the recent "Directory of a sparse Map" fix (see
Bug 5822) in the following two ways:

1. Increase the "sparse Map threshold" (ratio of number of local
   elements in the Directory Map, to number of local elements in the
   original noncontiguous Map) from 2 to 10.

2. Use FixedHashTable instead of HashTable.

Using a hash table has a memory cost as well as a performance cost
(for both construction and lookups).  2x is probably not sufficiently
sparse to justify switching from the array-based implementation to the
hash table - based implementation.  We can tune this factor more if
there is interest.  For now, I'll set it to 10x.

FixedHashTable has a lower setup cost and better lookup performance
than HashTable.  We can use FixedHashTable in this case, because we
know all the keys and their values in advance.  The greater
flexibility of HashTable is unnecessary here.

The second optimization required adding a new constructor to
FixedHashTable.  It accepts an arbitrary list of values, rather than
assuming that values occur in a contiguous sequence.  When Map uses
FixedHashTable for GID-to-LID lookups, the LIDs (the values in the
table) occur in a contiguous sequence, so there is no need to accept
an arbitrary sequence of values.  For this application in Directory,
FixedHashTable needs to take an arbitrary list of values.
FixedHashTable works fine in this case; it just needed a one-line
change to initialization.  For now, I've duplicated the initialization
code in FixedHashTable for both cases.  It would be nice sometime to
factor out the common code from the two initialization cases.