Tribits: Changes to support new dashboard features for coverage testing
David Cole had provided a set of changes as a start for being able to support a
coverage build that could distinguish between library and test code. This commit
applied the concepts from those commits such that the coverage build will work
without having some of the side effects that were undesirable. This required
changing what labels were attached to many of the targets that are built to
declare which part of the coverage they belong to, without breaking the
association to the package that it used to state. There was one major deviation
from the changes that David provided. The test names are not labeled with
<package_name>Test. This is because when they were labeled as such packages were
appearing on the dashboard labeled not with their name, but instead "2 labels"
since there was a <package_name> label and a <package_name>Test label for every
package with tests. This change does not affect the coverage testing because
test are cmake targets and not associated directly with source files for the
dashboard.

This also required making a few changes to the testing driver core to make sure
that the new labels were used for selecting what packages to build and to make
sure the correct set of information was submitted to the dashboard.

There is one caveat with this. To support subpackages and normal packages
cleanly normal packages end up with a redundant label for libraries. Subpackages
needed to be able to label their libraries with the subpackage's name, but to be
able to communicate to the dashboard correctly they also need to label their
libraries with the parent package's name. For Normal packages the package name
and parent package name are the same which leads to 2 identical labels being set
for normal packages. This doesn't have any visible side effects for the
dashboard, however.

For packages you will now see at least 2 labels on the coverage report.
Generally this will be "<package_name>Exes" and "<package_name>Libs" where the
Exes generally are the sources tests and Libs refers to sources compiled into a
library. For packages with subpackages there will be additional entries for each
subpackage.