Piro: Revised ObserverBase interface to avoid aliasing
Passed the stamp by value instead of passing it by (const) reference to
avoid potential aliasing, e.g. with the contents of the vector(s) passed
as input arguments.

This is a breaking change that will hopefully not be too disruptive
since no official release of Trilinos has occurred since the
introduction of Piro::ObserverBase.