Belos: Tpetra OrthoManager test: Map & CrsMatrix creation "generic."
I made creation of the Map and CrsMatrix in this test "generic."  The
latter word is in quotes because the Harwell-Boeing reader for Tpetra
sparse matrices currently only knows how to read matrices with
Scalar=double.  However, the test does not have to read a sparse
matrix, and the loader function also creates a Map for the test to
use, so the function can properly now be called generic.