Teuchos: Fix for (reopened) Bug 4457.
In particular, this fix addresses the "operator ArrayView<const T>()"
issue.  That is, almost all compilers but GCC warn on that instance
method of ArrayView, because if T = const X for some type X, then that
method will never get called (and doesn't make sense anyway, since
"const const X" is not a valid type).  The simplest and least
disruptive fix is to add a partial specialization of ArrayView for
const T.  I also considered the following fixes:

1. enable_if - style template metaprogramming, to enable the offending
   method only when it makes sense for the given type T.

2. Make both ArrayView<T> and ArrayView<const T> derive from a common
   base class ArrayViewBase<T>, and only include in ArrayView<T> (not
   in the ArrayView<const T> specialization) the offending methods.

Eric Phipps pointed out that enable_if would not work for a conversion
operator, because a conversion operator has neither a return type
(it's kind of like a constructor) nor arguments which can be used as
template parameters for enable_if.  Furthermore, an enable_if solution
would require that we introduce an enable_if implementation into
Teuchos, since Teuchos cannot have a required dependency on Boost's
implementation of enable_if.  I would prefer not to expand the public
interface of Teuchos if I can avoid it.

The second alternate solution is attractive since it reduces the
amount of duplicated code in the partial specialization vs. the
generic implementation.  However, it interferes with the circular
dependency between ArrayView and ArrayRCP.  (In a debug build,
ArrayView uses ArrayRCP to do bounds checking, whereas an ArrayRCP can
return a nonpersisting view as an ArrayView.  That's OK because only
the implementation of each of these two classes needs to include the
other class' implementation; it's possible to separate declaration
cyclicness from implementation cyclicness.  However, if ArrayView
derives from some ArrayViewBase class, then it has to include the
implementation of ArrayRCP.  There is probably a way to resolve this,
but I didn't feel like it was worth the complication.)

Dismissing these two options left me with the "partial specialization
with no base class" solution, which this commit implements.  Thanks to
Ross Bartlett, Carter Edwards, Brent Perschbacher, and Eric Phipps for
ideas and discussion.

Build/Test Cases Summary
Enabled Packages: Teuchos, Tpetra
0) MPI_DEBUG => passed: passed=165,notpassed=0 (1.47 min)
1) SERIAL_RELEASE => passed: passed=145,notpassed=0 (2.00 min)