Modify mesh interface to provide communication data and modify salinas and eigen to use it
Added fillElementSideCommunicationList and fillNodeCommunicationList functions which
return the element/side and node communication list data.
The element and node ids are 1-based processor local ids and the
side ordinal is 1-based.

Added getNumberProcessorsInDecomposition function which specifies how many processors the mesh
passed to salinas is running on which can be different than the number of processors Salinas is
using.

Added a FmwkOriginatedStkMeshInterface class derived from StkMeshInterface class
which is a stk mesh that comes from a framework-based app.  It takes a Fmwk::MeshBulkData
argument instead of a stk::mesh::BulkData argument.  It currently overrides three functions:
getNumberProcessorsInDecomposition
fillElementSideCommunicationList
fillNodeCommunicationList

In addition, the selector it uses to select locally-owned and globally-shared parts is
based on the framework bulk data.

The class and functions in it should go away as more framework-isms are removed.  It provides
a visible reminder of what differences exist between instances of stk-mesh use.

Modified the salinas eigen interface to not have to know about the various mesh interface
implementations and it now takes a sierra::Mesh* instead of taking the stk or other arguments
and creating the mesh internally -- makes it easier to add new mesh interface subclasses without
modifying downstream code

No more nemesis get calls or ex_get related to non-transient data although the mesh file must still
exist due to some other calls that Salinas makes; will try to remove those if possible so an eigen analysis
can run completly off of an in-memory mesh.

(From Sierra)